Update Login Authentication Strategy
Previous Strategy: Allow frontend to communicate with Google, retrieve profile information and token, then return to backend.

Current Strategy: User clicks a button which navigates to a server route (/auth/google), does the info transfer with google, then redirects back to the client login route with a token stored in the cookies (to verify the user against the server for future requests)  and the user UUID stored in the route. From here, the frontend then parses the information and makes another call to the API for more user information. User data is then stored in local storage.