Update APIHandler to remove interface
- Instead of a APIHandler with 2 swappable classes of Fetcher and MockFetcher, revise to have 2 classes: APIHandler and MockAPIHandler.

The original interface design was because I thought it would be better for the rest of the codebase to always interact with one class: APIHandler. As opposed to switching between 2 classes depending on whether it was a development or testing environment. However, after looking at the kind of tests I am currently writing, I am not doing the kind of integration tests that requires the rest of the codebase to only know one class. I am injecting the Fetchers directly into each component as a prop.