Fix Date formatting bug
Issue: After updating an item, the date thats sent to the server is always 1 day before the Date originally sent by the server. This is strange as the date was supposed to be formatted on incoming and outcoming, and it should not be changed at all. After a closer look, the issue has to do with the`dayEndedUTCString` function -- it needed the passed in string format to be "07-01-2019" to work correctly. However, date strings from a html input comes in as "2019-07-01" which makes the whole function not work as expected. The fix is to run the date through the `date-fns.parse` function before converting it to a ISOString.