Fix Form submission bug
Issue: When a user fills out the form and clicks submit, if there is an error, the form will continously try and re-submit the request, making infinite calls to the server. Another issue that was noticed is that, after submitting a successul form and if the component is still mounted (aka, we weren't redirected after a succcessful submit, and user is still on the same page and can interact with the form), then any change to the input (like changing the values) will trigger another submit.

Fix: The issue had to do with useEffect in useForm. We just had to toggle "isSubmitting" back to false after error/success changes.