Fix item won't update after first bid bug
Issue: When creating a new bid on an item, the validation and subtext on the form field doesn't update to show the new minimum price.

Fix: API sends 2 types of prices, price for the listing price, and current_highest_bid to keep track of the bid prices on the item. When there are no bids, the current_highest_bid was set to be the same as price -- which makes it very difficult on the frontend to determine whether a bid has been made and whether the minimum price should be updated. We updated the API to set current_highest_bid to nil  and use conditionals to determine the current item price and minimum price necessary.