Fix second bid submission bug
Issue: When submitting a bid, it will re-render the page. However, a second consecutive bid submission will not trigger a re-render of the Item / Bids.

Fix: This is because a successful bid triggers a fetchItems call, and then never resets "refetch" back to false , so subsequent submissions does not know the page has changed enough to trigger the re-render.  We just had to make sure that refetch is set to false after the useEffect fetchItems call